refactor : DisplayedName 공통 model 리팩토링
